Solve Homework1&2
This week was dedicated to the on-chain smart contract validation for which we
needed to understand the advanced concept of typed instances within Haskell and
templating to acommodate type-safty for our plutus validators.

For the first homework assignment we could simply resort to pattern matching on
a tuple of booleans as it is already pre-declared

See: https://youtu.be/E5KRk5y9KjQ?t=3511

Regretfully I do not understand from the source code how that exactly happens:
https://github.com/input-output-hk/plutus/blob/master/plutus-tx/src/PlutusTx/IsData/Class.hs#L49-L54

For the second homework assignment we needed to define our own redeemer type which
in this case was a record type.